requirements.txt
# Using Python 3.11.8
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-webpack-loader==3.1.0
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.3.0
pycparser==2.22
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1